Fix (plan) output when updating k8s addons
For the commands `utils update-aws-node` and `utils update-coredns` when
used in dry-run mode, it now returns `already up-to-date` if the image
tag of the deployment/daemonset matches what is in the built-in assets.

While this does _not_ confirm a run with `--approve` is a no-op (since the
current image might not be region-aware or other parts in the manifest
differ), it at least tells you that the version is up-to-date.

When running the commands with `--approve`, the built-in manifests will be
applied idempotently in any case no matter whether the image tags match
or not.